+++ openrtm_tools/catkin.cmake
@@ -8,12 +8,13 @@ find_package(catkin REQUIRED COMPONENTS rostest)
 
 ## System dependencies are found with CMake's conventions
 # find_package(Boost REQUIRED COMPONENTS system)
+catkin_add_env_hooks(99.rtmlaunch SHELLS sh DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/env-hooks)
 
 
 ## Uncomment this if the package has a setup.py. This macro ensures
 ## modules and global scripts declared therein get installed
 ## See http://ros.org/doc/api/catkin/html/user_guide/setup_dot_py.html
-# catkin_python_setup()
+catkin_python_setup()
 
 #######################################
 ## Declare ROS messages and services ##
@@ -120,7 +121,9 @@ include_directories(
 #   DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
 # )
 
+
 install(DIRECTORY scripts test DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION} USE_SOURCE_PERMISSIONS PATTERN ".svn" EXCLUDE)
+install(PROGRAMS scripts/rtmlaunch scripts/rtmtest DESTINATION ${CATKIN_GLOBAL_BIN_DESTINATION})
 
 #############
 ## Testing ##
